Conflicts:
	CMakeLists.txt
	ECS/Any.hpp
	Threads/SafeFifo.hpp
	Threads/UMutex.cpp
	components/Box2DComponent.cpp
	components/CMakeLists.txt
	components/Friction2DComponent.cpp
	components/MoveFollowComponent.cpp
	components/MoveFollowComponent.hh
	components/Pos2DComponent.cpp
	components/SFMLSpriteComponent.cpp
	components/SFMLSpriteComponent.hh
	components/Speed2DComponent.cpp
	main_client.cpp
	main_server.cpp
	systems/CMakeLists.txt
	utility/ComponentFactory.hpp

+++ Network/NetworkBuffer.cpp
@@ -233,7 +233,7 @@ void		NetworkBuffer::setOffset(unsigned int offset)
   int		diff;
 
   diff = offset - this->_offset;
-  if (this->_buffer_size - diff >= 0)
+  if ((int)this->_buffer_size - diff >= 0)
     {
       this->_buffer += diff;
       this->_buffer_size -= diff;
