+++ Network/NetworkBuffer.hh
@@ -75,7 +75,7 @@ private:
       i = 0;
     else
       i = sizeof(T) - 1;
-    while ((isBigEndian() && i < sizeof(T)) || (!isBigEndian() && i >= 0))
+    while ((isBigEndian() && i < static_cast<int>(sizeof(T))) || (!isBigEndian() && i >= 0))
       {
 	this->_buffer[this->_current_pos] = tab[i];
 	if (isBigEndian())
@@ -102,7 +102,7 @@ private:
       i = 0;
     else
       i = sizeof(T) - 1;
-    while ((isBigEndian() && i < sizeof(T)) || (!isBigEndian() && i >= 0))
+    while ((isBigEndian() && i < static_cast<int>(sizeof(T))) || (!isBigEndian() && i >= 0))
       {
 	tab[i] = this->_buffer[this->_current_pos];
 	if (isBigEndian())
